:recycle: refactor(web/admin/client): Moved the image update logic out of the ImageEditor component as much as possible
Provide properties for customization:
1. onChange - input change handler
2. handler1 - first button click handler
3. handler2 [optional] - second button click handler
4. button1Text - text for the first button
5. button2Text [optional] - text for the second button

+++ src/LiWiMus.Web.Admin/client/src/pages/PlaylistDetails.page.tsx
@@ -9,17 +9,29 @@ import playlistCover from "../images/playlist-cover-negative.png";
 import { useSnackbar } from "notistack";
 import axios from "../services/Axios";
 
+const API_URL = process.env.REACT_APP_API_URL;
+
 export default function PlaylistDetailsPage() {
   const { id } = useParams() as { id: string };
-  const [playlist, setPlaylist] = useState<Playlist | null>();
+  const [playlist, setPlaylist] = useState<Playlist>();
+  const [playlistPhoto, setPlaylistPhoto] = useState<string>("");
   const [loading, setLoading] = useState(true);
   const { enqueueSnackbar } = useSnackbar();
+
+  const setPlaylistWithPhoto = (playlist: Playlist) => {
+    const photoLocation = playlist.photoLocation
+      ? API_URL + playlist.photoLocation
+      : playlistCover;
+    setPlaylist({ ...playlist });
+    setPlaylistPhoto(photoLocation);
+  };
 
   useEffect(() => {
     axios
       .get(`/playlists/${id}`)
       .then((response) => {
-        setPlaylist(response.data as Playlist);
+        const playlist = response.data as Playlist;
+        setPlaylistWithPhoto(playlist);
       })
       .catch((error) => enqueueSnackbar(error.message, { variant: "error" }))
       .then(() => setLoading(false));
@@ -33,8 +45,35 @@ export default function PlaylistDetailsPage() {
     return <NotFound />;
   }
 
-  const updatePlaylistPhoto = async (photo: File) => {
+  const updatePhotoHandler = (input: HTMLInputElement) => {
+    input.click();
+  };
+
+  const removePhotoHandler = async () => {
+    if (!playlist.photoLocation) {
+      return;
+    }
     try {
+      await axios.delete(playlist.photoLocation);
+      const response = await axios.post(`/playlists/${id}/removePhoto`);
+      setPlaylistWithPhoto(response.data as Playlist);
+      enqueueSnackbar("Photo removed", { variant: "success" });
+    } catch (error) {
+      // @ts-ignore
+      enqueueSnackbar(error.message, { variant: "error" });
+    }
+  };
+
+  const changePhotoHandler = async (
+    event: React.ChangeEvent<HTMLInputElement>
+  ) => {
+    const input = event.target;
+    if (!input.files || !input.files[0]) {
+      return;
+    }
+    try {
+      const photo = input.files[0];
+
       if (playlist.photoLocation) {
         await axios.delete(playlist.photoLocation);
       }
@@ -47,29 +86,12 @@ export default function PlaylistDetailsPage() {
       const photoLocation = data.location as string;
       const updateDto = { id, photoLocation };
       const response = await axios.patch("/playlists", updateDto);
-      setPlaylist(response.data as Playlist);
+      setPlaylistWithPhoto(response.data as Playlist);
       enqueueSnackbar("Photo updated", { variant: "success" });
     } catch (error) {
       // @ts-ignore
       enqueueSnackbar(error.message, { variant: "error" });
     }
-    return playlist;
-  };
-
-  const removePlaylistPhoto = async () => {
-    if (!playlist.photoLocation) {
-      return playlist;
-    }
-    try {
-      await axios.delete(playlist.photoLocation);
-      const response = await axios.post(`/playlists/${id}/removePhoto`);
-      setPlaylist(response.data as Playlist);
-      enqueueSnackbar("Photo removed", { variant: "success" });
-    } catch (error) {
-      // @ts-ignore
-      enqueueSnackbar(error.message, { variant: "error" });
-    }
-    return playlist;
   };
 
   return (
@@ -88,10 +110,10 @@ export default function PlaylistDetailsPage() {
         >
           <ImageEditor
             width={250}
-            src={playlist.photoLocation}
-            imagePlaceholderSrc={playlistCover}
-            updatePhoto={updatePlaylistPhoto}
-            removePhoto={removePlaylistPhoto}
+            src={playlistPhoto}
+            onChange={changePhotoHandler}
+            handler1={updatePhotoHandler}
+            handler2={removePhotoHandler}
           />
           <h1>{playlist.name}</h1>
         </Grid>
